Tested staff processes

+++ src/main/java/lk/ijse/cmjd108/LibMgmt/controller/StaffController.java
@@ -1,6 +1,7 @@
 package lk.ijse.cmjd108.LibMgmt.controller;
 
 import lk.ijse.cmjd108.LibMgmt.dto.StaffDto;
+import lk.ijse.cmjd108.LibMgmt.exception.StaffNotFoundException;
 import lk.ijse.cmjd108.LibMgmt.service.StaffService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -27,19 +28,52 @@ public class StaffController {
 
     @DeleteMapping
     public ResponseEntity<Void> deleteStaff(@RequestParam("staffId") String staffId){
-        staffService.deleteStaff(staffId);
-        return ResponseEntity.noContent().build();
+        if(staffId == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            staffService.deleteStaff(staffId);
+            return ResponseEntity.noContent().build();
+        }catch (StaffNotFoundException e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     @PatchMapping(value = "/{staffId}",consumes = MediaType.APPLICATION_JSON_VALUE )
     public ResponseEntity<Void> updateStaff(@PathVariable String staffId, @RequestBody StaffDto staffDto){
-        staffService.updateStaff(staffId,staffDto);
-        return ResponseEntity.noContent().build();
+        if(staffId == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }try {
+            staffService.updateStaff(staffId,staffDto);
+            return ResponseEntity.noContent().build();
+        }catch (StaffNotFoundException e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     @GetMapping("{staffId}")
-    public ResponseEntity<StaffDto> getSelectedStaff(@PathVariable String staffId){
-        return ResponseEntity.ok(staffService.getSelectedStaff(staffId));
+    public ResponseEntity<StaffDto> getSelectedStaff(@PathVariable String staffId, @RequestBody StaffDto staffDto){
+        if(staffId == null || staffDto == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }try {
+            staffService.updateStaff(staffId,staffDto);
+            return ResponseEntity.noContent().build();
+        }catch (StaffNotFoundException e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     @GetMapping
